% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CV.RBDmodel.R
\name{CV.RBDmodel}
\alias{CV.RBDmodel}
\title{CV.RBDmodel}
\usage{
\method{CV}{RBDmodel}(model, n = 5)
}
\arguments{
\item{model}{RBDmodel object}

\item{n}{int(= 5) defining number of Cross Validations}
}
\value{
list of class RBDerr with 3 elements
\itemize{
    \item RBDparams list of cross validation parameters
    \item RBDerr matrix with varying rows and 4 columns: Rows represent a classification each. 
          Columns represent prediction errors: "TP", "FP", "TN", "FN") 
    \item RBDpp num arr of posterior probabilities of each protein to be RBP 
 }
}
\description{
Estimate prediction error of a RBDmodel using n-fold Cross Validation.
}
\details{
5-fold Cross Validation is a common method to estimate prediction errors.
This function can run for a while (roughly n-times as long as the corresponding RBDfit).
}
\examples{
l <- c("A", "E", "K", "S", "R", "Q", "L", "T", "N", "G", "M", "W", "D", "H", "F", "Y", "C", "I", "P", "V")
sequences <- sapply(sample(10:100, 500, replace = TRUE), function(x) paste(sample(l, x, replace = TRUE), collapse = ""))
labels <- c(rep(1, 200), rep(0, 300))

model <- RBDfit(sequences, labels, k = 2)

err <- CV(model)
str(err)
err
plotROC(err)
plot(err)

}

