% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvRBDfit.R
\name{cvRBDfit}
\alias{cvRBDfit}
\title{cvRBDfit}
\usage{
cvRBDfit(model, seq, labels, nfold = 5)
}
\arguments{
\item{model}{RBDmodel created with RBDfit, this is the model for which to estimate prediction error}

\item{seq}{chr arr containing amino acid sequences of proteins}

\item{labels}{int arr containing labels: 0 = NoRBP, 1 = RBP}

\item{nfold}{int(= 5) defining number of Cross Validations}
}
\value{
list of class RBDerr with 3 elements
\itemize{
    \item RBDparams list of cross validation parameters
    \item RBDerr matrix with varying rows and 4 columns: Rows represent a classification each. 
          Columns represent prediction errors: "TP", "FP", "TN", "FN") 
    \item RBDpp num arr of posterior probabilities of each protein to be RBP 
 }
}
\description{
Estimate prediction error of a RBDfit using n-fold Cross Validation.
}
\details{
5-fold Cross Validation is a common method to estimate prediction errors.
This function can run for a while (roughly n-times as long as the corresponding RBDfit).
}
\examples{
l <- c("A", "E", "K", "S", "R", "Q", "L", "T", "N", "G", "M", "W", "D", "H", "F", "Y", "C", "I", "P", "V")
sequences <- sapply(sample(10:100, 500, replace = TRUE), function(x) paste(sample(l, x, replace = TRUE), collapse = ""))
labels <- c(rep(1, 200), rep(0, 300))

model <- RBDfit(sequences, labels, k = 2)
err <- cvRBDfit(model, sequences, labels, nfold = 3)

str(err)
err
ROCplot(err, sens = 0.5)
plot(err)

}

