% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Err.R
\name{Err}
\alias{Err}
\title{Err}
\usage{
Err(posterior, labels, n = 10000, equal = "freq")
}
\arguments{
\item{posterior}{num arr of posterior probabilities of each datapoint}

\item{labels}{int arr of factor of length(posterior) of true class labels. must have {0, 1}}

\item{n}{int(= 10000) for how many thresholds will be set between 0 and 1 to compute matrices.
For equal="freq" this is an estimation}

\item{equal}{chr(= freq) "width" or "freq". set thresholds with equal width or equal frequency}
}
\value{
matrix of n row and 4 cols with TP, FP, TN, FN
}
\description{
compute confusion matrices from posterior probabilities and labels for a number of thresholds
thresholds will be distributed between 0 and 1. posterior > t will assign class as 1
}

