% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetKMers.R
\name{GetKMers}
\alias{GetKMers}
\title{GetKMers}
\usage{
GetKMers(seq, K = 3)
}
\arguments{
\item{seq}{character as input sequence}

\item{K}{int defining length of K-mer}
}
\value{
character arr of all K-mers found in seq
}
\description{
Extract all K-mers of a character sequence.
}
\examples{
GetKMers(paste(letters[1:10], collapse = ""))

}

